// src/prisma/generated/models/User.ts
export type UserWhereInput = {
  AND?: Prisma.UserWhereInput | Prisma.UserWhereInput[];
  OR?: Prisma.UserWhereInput[];
  NOT?: Prisma.UserWhereInput | Prisma.UserWhereInput[];
  id?: Prisma.IntFilter<'User'> | number;
  name?: Prisma.StringFilter<'User'> | string;
  email?: Prisma.StringFilter<'User'> | string;
  password?: Prisma.StringFilter<'User'> | string;
  role?: Prisma.EnumRoleFilter<'User'> | $Enums.Role;
  createdAt?: Prisma.DateTimeFilter<'User'> | Date | string;
  updatedAt?: Prisma.DateTimeFilter<'User'> | Date | string;
};

export type UserOrderByWithRelationInput = {
  id?: Prisma.SortOrder;
  name?: Prisma.SortOrder;
  email?: Prisma.SortOrder;
  password?: Prisma.SortOrder;
  role?: Prisma.SortOrder;
  createdAt?: Prisma.SortOrder;
  updatedAt?: Prisma.SortOrder;
};